sua tim kiem tour

+++ src/main/java/com/app/controller/TourController.java
@@ -34,7 +34,7 @@ public class TourController {
 	
 	@RequestMapping("/list/{page}")
 	public String showProduct(Model model, @ModelAttribute("searchForm") PagingSearchFilterProduct searchForm, HttpSession session, 
-			@PathVariable("page") int page, @RequestParam("provinceId") int provinceId) {
+			@PathVariable("page") int page, @RequestParam("provinceId") Long provinceId) {
 		searchForm.setPage(page);
 		searchForm.setProvinceId(provinceId);
 		Page<Product> pageProduct = productService.getAll(searchForm);
